chore: add type directory
