+++ src/api/youtube/index.ts
@@ -2,22 +2,6 @@ import { QueryFunctionContext } from "@tanstack/react-query";
 
 import { youtubeApi } from "@/lib/axios";
 
-export interface YoutubeSearchSnippet {
-  title: string;
-  thumbnails: {
-    medium: {
-      url: string;
-    };
-  };
-}
-
-export interface YoutubeSearchItem {
-  id: {
-    videoId: string;
-  };
-  snippet: YoutubeSearchSnippet;
-}
-
 export const getYoutubeSearch = async ({ queryKey }: QueryFunctionContext) => {
   const [_, keyword] = queryKey;
   const res = await youtubeApi.get("", {
