+++ src/type/tab/index.ts
@@ -1,0 +1,13 @@
+export type TabInfoType = {
+  id: string;
+  title: string;
+  artist: string;
+  thumbnailUrl: string;
+};
+
+export interface TabApiResponse {
+  id: string;
+  title: string;
+  artist: string;
+  thumbnail_url: string;
+}
