+++ src/type/youtube/index.ts
@@ -1,0 +1,15 @@
+export interface YoutubeSearchSnippet {
+  title: string;
+  thumbnails: {
+    medium: {
+      url: string;
+    };
+  };
+}
+
+export interface YoutubeSearchItem {
+  id: {
+    videoId: string;
+  };
+  snippet: YoutubeSearchSnippet;
+}
